requirements.txt
PyYAML==5.4.1
requests==2.26.0
urllib3==1.26.6
click==8.0.1
